+++ ahlive/tests/test_data.py
@@ -237,7 +237,7 @@ def test_config_bar_chart(preset):
         assert (actual == expected).all()
     else:
         actual = ds["x"].values.ravel()
-        expected = [1, 1, 1, 2, 2, 2]
+        expected = [0, 0, 0, 1, 1, 1]
         assert (actual == expected).all()
 
     actual = ds["y"].values.ravel()
@@ -820,16 +820,56 @@ def test_config_wave_chart(how):
     y1 = [4, 5, 6, 7, 8]
     y2 = [8, 4, 2, 3, 4]
     ah_obj = (
-        ah.Array(x, y1, preset="morph", chart="line", label="A", group="C")
-        * ah.Array(x, y2, chart="line", label="A", group="C")
-        * ah.Array(x, y2, label="B", group="C")
+        ah.Array(x, y1, preset="morph", chart="line", group="A")
+        * ah.Array(x, y2, chart="line", group="A")
+        * ah.Array(x, y2, group="B")
     )
     if how == "even":
-        ah_obj *= ah.Array(x, y1, chart="line", label="B", group="C")
+        ah_obj *= ah.Array(x, y1, chart="line", group="B")
     ah_obj = ah_obj.finalize()
     ds = ah_obj[1, 1]
     assert len(ds["item"] == 2)
     assert len(ds["batch"] == 5)
     assert len(ds["state"] == 30)
-    assert (ds["label"].values == ["A", "B"]).all()
-    assert (ds["group"].values == ["C", "C"]).all()
+    assert (ds["group"].values == ["A", "B"]).all()
+
+
+@pytest.mark.parametrize("dtype", ["numeric", "datetime"])
+@pytest.mark.parametrize("chart", ["line", "scatter"])
+@pytest.mark.parametrize("first", [False, True])
+def test_remark(dtype, chart, first):
+    if dtype == "numeric":
+        x = np.array([0.0, 1, 1, 3])
+        xs_condition = [0, 1]
+    elif dtype == "datetime":
+        x = pd.to_datetime(
+            ["2017-02-01", "2017-02-02", "2017-02-02", "2017-02-03"]
+        ).values
+        xs_condition = pd.to_datetime(["2017-02-01", "2017-02-02"])
+    y = [4, 5, 6, 7]
+
+    ah_obj = (
+        ah.Array(x, y, chart=chart).remark("x", xs=xs_condition).remark("abcdef", ys=7)
+    )
+    ds = ah_obj[1, 1]
+    remarks = ds["remark"].squeeze().values
+    assert ds["remark"].ndim == 2
+    if not first:
+        assert (remarks[:3] == x[:3].astype(str)).all()
+        assert remarks[-1] == "abcdef"
+    else:
+        assert (remarks[:2] == x[:2].astype(str)).all()
+        assert remarks[-1] == "abcdef"
+
+
+def test_remark_overlay():
+    x = np.array([0, 1, 1, 3])
+    y1 = [4, 5, 6, 7]
+    y2 = [8, 9, 10, 11]
+
+    ah_obj = ah.Array(x, y1) * ah.Array(x, y2)
+    ah_obj = ah_obj.remark("abc", ys=[4, 11])
+    ds = ah_obj[1, 1]
+    remarks = ds["remark"]
+    assert remarks.isel(item=0, state=0).item() == "abc"
+    assert remarks.isel(item=1, state=-1).item() == "abc"