+++ CHANGELOG.rst
@@ -68,6 +68,9 @@ bug fixes
 - Move `_config_chart` before `_precompute_base` to fix automated preset text format (:pull:`76`).
   By `Andrew Huang <https://github.com/ahuang11>`_
 
+- Fix various issues: formatting, ease of use, use of `group` instead of `label` for `morph` preset, negative bar values, compressing vars, bar width, refactor of remarks, and interpolation of bar labels. (:pull:`76`).
+  By `Andrew Huang <https://github.com/ahuang11>`_
+
 internal changes
 ~~~~~~~~~~~~~~~~